requirements.txt
django==4.2.1
python-dotenv==1.0.0
djangorestframework==3.14.0
confluent-kafka==2.1.1
